Update patch set 1
Patch Set 1:

I like this idea!

We have a ton of death tests, but to date, nary a no-compile test. It'd be worthwhile to set up infrastructure for that to test the static_assert here. I'd have to look at the .nc infrastructure from Chrome to see if we could plug it in here.

Patch-set: 1
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>