Update patch set 7
Patch Set 7:

> I've implemented an identical interface, with the slight hack of the UserMinidumpStream still exposing the raw stream data as a 'MemorySnapshot'. The only effect of this is that the MemorySnapshot returned has an address of 0, which seems more desirable than implementing the method with another interface.

I don’t agree. This seems like a hack. It’s not a MemorySnapshot, we shouldn’t pretend that it is.

Even if it weren’t for the address field, this kind of coding negatively impacts readability. If future engineers can’t trust that a MemorySnapshot is a snapshot of memory as it advertises, what have we really gained?

Patch-set: 7
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>