Update patch set 2
Patch Set 2: Code-Review+1

Unfortunately this "lock" class doesn't do anything on Fuchsia other than that DCHECK. Relooking at https://chromium-review.googlesource.com/924312, the plan was that we never use this in production because there was work in progress for advisory locks at the file system level, but that sadly has not panned out. (There was some CLs and meetings, but I can't find the bugs now, ZX-264 LE-437 are similar.)

The main argument boiled down to that each app on Fuchsia would have a separate private data store, so it was unnecessary to implement file advisory locks because the data would have only one mutator. I don't believe that's true today in Fuchsia unfortunately because the crash database is stored in /data, so e.g. both the crash reporter and a command line tool could be changing the database at the same time, requiring a lock.

Still it doesn't make sense to leave the lockfile lying around in release-only, so lgtm to fix the side-effect-in-DCHECK.

Patch-set: 2
CC: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1