Update patch set 2
Patch Set 2: Code-Review+1

OK, then LGTM to this as-is. But consider whether we need to make a follow-up request to Fuchsia to be able to handle this in a more deterministic and synchronous way. It doesn’t seem like an extraordinary request if the system’s already guaranteeing that no _user_ code will run once the call returns.

Patch-set: 2
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1