Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> After seeing https://chromium-review.googlesource.com/c/chromium/src/+/1390906, I have to wonder:
> 
> Why was a custom ArraySize added to Crashpad instead of using base::size, which seems to be the new standard in Chromium?

We're going to roll base::size into being std::size when we get C++17, and as a module in third_party I wanted to avoid doing this again in the future. There are also tons of places where base::size doesn't work and we have to use std::extent, and because you already had a macro it seemed reasonable to move you to a macro and avoid the hassle.

Patch-set: 5