Update patch set 2
Abandoned

The source of the build problem downstream was diagnosed as being a problem with the downstream’s build description, and not with any compiler’s failure to interpret #include paths correctly. The downstream build description was revised, and this change is no longer necessary.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon