Update patch set 4
Patch Set 4: Commit-Queue+1

I took your patch on top of mine and the CQ’s into it.

The build from a macOS host was still busted, fixed at https://chromium-review.googlesource.com/c/1506239/. I don’t want to bring that DEPS roll into this change because it’ll conflict with the other DEPS roll I want to make (https://chromium-review.googlesource.com/c/1503686) and resolving that cleanly wouldn’t actually be so clean because it’d turn the Fuchsia builders red temporarily. Everything here is codependent in a weird way.

So the plan is to land this, which lets the Fuchsia builders stay green, and let the not-so-important-because-it’s-not-tested Mac-to-Fuchsia build break momentarily, then land my -Wextra-semi enabler, and finally do a DEPS roll to fix the Mac-Fuchsia cross build.

Patch-set: 4
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1