Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> so if we run `crasher write0_mt` they go away right? and if you were to build with `--monolith garnet/packages/config/crashpad_upload_to_prod_server` we would upload something on go/crash for the other threads in the crash process?

Yes!

However I just realized that there's a single new log in its place:

[00141.296157][10645][10657][klog] INFO: [10645:10657:20190322,211641.282431:ERROR scoped_task_suspend.cc:38] thread failed to suspend: ZX_ERR_TIMED_OUT (-21)

That happens even when doing a single threaded crash, so I think it's that the already-suspended thread doesn't re-assert that it got suspended.

It does fix the stacks though when uploading; crash/6099831c532b391f is an example of `crasher write0_mt`.

I'll try to track down that new log, otherwise I might land this as-is as I'm not sure if perhaps that should be fixed on the other side of the fence.

Patch-set: 5
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>