Update patch set 3
Patch Set 3: Code-Review+1

LGTM.

I thought that this was referenced in more locations, but since it’s not, this is the right move. We can always factor it out of test.gni if we do need to access it from elsewhere.

Patch-set: 3
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1