Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

I'm concerned that we're going to end up with confused stacks on the crash server, for anything that crashes in an executable because they'll all be named <_> now. The bucketing code might merge things that shouldn't be merged either.

That'll only affect testing/dev builds not "Final" builds, but it'd also happen if something breaks in the symbol upload pipeline.

Anyway, lgtm to not suffer this change any more, but we'll have to watch and see how it goes. Most crashes appear to be in .so files anyway.

Could you also investigate if a unittest could be added to process_reader_fuchsia_test.cc?

Patch-set: 3
Reviewer: Gerrit User 1144260 <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1