ProcessSnapshot::CommandLine and ProcessSnapshot::Environment
Continuation of
https://chromium-review.googlesource.com/c/crashpad/crashpad/+/465827

This is a crashpad part of a restoration of the 'chrome crashed' dialog.
The dialog was lost during the breakpad->crashpad migration.

Previously the chrome main process restarted itself after a minidump upload. Such an approach seems to be wrong since it is not safe to do things in a crashed process.
It seems better to restart the main process from the crashpad handler. To do so the crashpad process needs to know command-line and environment of the crashed process.

Bug: crbug.com/602956
Change-Id: Ida2e9cd471d794816d1cf1951961e11c86eb37b5