Enable fuzz target to build in Chromium.
Enable building elf_image_reader_fuzzer in Chromium.
Rename it to crashpad_elf_image_reader_fuzzer so that its clearer
where the fuzzer comes from.
Import chromium's fuzzer_test definition when building in Chromium and
create an empty fuzzer group as is often necessary for fuzzers to be
built in Chromium.

Bug: 950093
Change-Id: I8afc104d26871311b04931b82a1600614a81bfc8