[fidl] Convert from - to . names for //zircon/public/fidl
This is part of a soft transition to replace fuchsia-foo-bar names
with fuchsia.foo.bar names for the directory and GN target names for
FIDL libraries defined in //zircon/system/fidl and accessed in legacy
GN via //zircon/public/fidl.  After the transition, the directory and
GN target names will exactly match the FIDL library names everywhere.
 1. Make //zircon/public/fidl/a.b.c alias //zircon/public/fidl/a-b-c
 2. Convert all //zircon/public/fidl uses outside fuchsia.git & integrate
 3. Convert all //zircon/public/fidl uses inside fuchsia.git
 4. Remove transition hacks
 5. Enforce naming conventions in GN templates
This is step 2.

Bug: ZX-3365 #comment //third_party/crashpad references converted
Change-Id: Ief90c6f9fa1da51dea7dbbd5b58741cbb7e06891