Update patch set 5
Patch Set 5:

Based on my read of that other code review, it’s not all aarch64 brk instructions that result in an illegal instruction exception, but the value that clang is using for __debugbreak() does result in this exception.

The change’s description, and the in-line comments that you’ve written, should be clearer on this point. The problem is the specific brk that clang uses to implement __debugbreak(), not Windows’ handling of brk.

Patch-set: 5
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>