Fix some debugbreak related tests on Windows ARM64
The specific brk instruction that clang uses to implement __debugbreak,
that cannot be recognized by Windows' exception handler
which causes an illegal instruction exception.
This CL is intent to modify the expectation value
at some point in these tests, until it is fixed in clang to emit
appropriate break point instruction for that.

Change-Id: I0673bbc816ccb1b6234b506592de29d145334709