Update patch set 1
Patch Set 1:

I've never heard of this CPU before, but the Internet seems to confirm that it's from something part-AMD-owned manufacturing in China.

Did you run tests? I think we have a test somewhere that verifies that it sees either GenuineIntel or AuthenticAMD when running on an x86-family CPU.

Patch-set: 1
CC: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>