client_win: Switch non-standard Rel Load/Acq Store -> Rel Store/Acq Load
client_win used a non-standard/poorly defined "Acquire Store" and "Release
Load" pair to have handlers notify when they changed state and to wait
for handlers to change state. Acquire stores and Release Loads are not
expressable in C++11 atomics and even taken at face value did not provide
useful semantics here (code waiting for a handler to change state wants to
see the handler's stores, not to ensure that the handler sees prior stores
of the waiter).

Switch to Release_Store / Acquire_Load.