Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Mark, thanks for your response! Moved the patch to crashpad, the all-in-one patch is
> 
> https://chromium-review.googlesource.com/c/chromium/src/+/1687090

I know you're not directly asking me to review this and Mark may have a different opinion, but personally I think it'd be helpful if you uploaded a series of dependent patches that include all the parts in Crashpad itself, instead of just this first piece. It is helpful to break CLs up into logical chunks like this, but only seeing the first one (where the code is not used other than in tests) isn't as clear. I realise you do have the all-in-one patch back in Chromium but probably still better to do it here where it's actually going to land :)

Patch-set: 1