Fix the build with an old NDK and a new API level.
It's occasionally useful to build with the API level set to a value beyond
what the NDK supports, for example when building with an old NDK (e.g. r16
which Chromium uses) and a custom compiler that activates features based on
the API level.

The android_get_device_api_level() function was added to the NDK in r20,
so don't try to use it without declaring it before then.

Change-Id: If58542c9d6b5315f823f2509f168f2cb79141e3f