Update patch set 3
Patch Set 3:

This is likely to have a large impact on the processor’s ability to reconstruct a backtrace beyond a handful of frames.

Are the stacks currently being captured viable for backtraces, or is it a bunch of garbage? You should be able to tell what range is necessary for a backtrace with a bit of minidump_stackwalk use.

Patch-set: 3
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>