Update patch set 2
Patch Set 2:

That seems to be an abusive use of sysroots. At the very least, it argues for a “we’d like to add this check, but Chromium infra is kind of broken, and we need to support glibc 2.xx”.

However, you still should be able to scope this to be glibc-only, right?

Patch-set: 2
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>