Update patch set 1
Patch Set 1:

The comment in ChromeCrashLoggedInDirect says that test can be removed when the new test ChromeCrashLoggedIn is stable. Does that mean that this option will no longer be necessary when that happens? This is only needed to make the legacy tests work?

https://cs.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/ui/chrome_crash_logged_in_direct.go?rcl=4985915f07c0541ee1bb596d70b8cec0cdcf7815&l=28

Patch-set: 1
Reviewer: Gerrit User 1172779 <1172779@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>