Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > The comment in ChromeCrashLoggedInDirect says that test can be removed when the new test ChromeCrashLoggedIn is stable. Does that mean that this option will no longer be necessary when that happens? This is only needed to make the legacy tests work?
> > 
> > https://cs.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/ui/chrome_crash_logged_in_direct.go?rcl=4985915f07c0541ee1bb596d70b8cec0cdcf7815&l=28
> 
> As the person who wrote that comment: I believe this option will be used in tast tests for the foreseeable future. It's not about those tests in particular, it's about crashes in other unrelated tests. We need to make sure test failures caused by Chrome crashes produce core / dmp files so that people can debug the failures. 
> 
> (As an aside, I'm not sure I agree with my past self. We should test the method that tast tests are using to gather core / dmp files, because they are important, even if they are not used on machines 'in the wild'. I may remove one of the two tests and leave the other.)

Thanks. Why do we need to write the dumps to a different location? Is that to avoid losing them to the cryptohome? I'm wondering if it'd be more self-contained to just pass on a --test-mode flag to crash_reporter and let it figure out the right place to write dumps when testing since it's already responsible for figuring the right dump locations in normal usage. (I also don't think the "unattended" naming is intuitive; I've had to remind myself what that means a couple times.)

Patch-set: 1
Reviewer: Gerrit User 1172779 <1172779@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>