Update patch set 3
Patch Set 3:

Now that I write all that... it occurs to me that perhaps this doesn't even _need_ to be in Crashpad.

Since it's now evolved to just using util/file/string_file.h, it could live in the Fuchsia tree? I not fully quite clear on how you're planning on factoring and where it will be called from though.

Patch-set: 3