Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Now that I write all that... it occurs to me that perhaps this doesn't even _need_ to be in Crashpad.
> > 
> > Since it's now evolved to just using util/file/string_file.h, it could live in the Fuchsia tree? I not fully quite clear on how you're planning on factoring and where it will be called from though.
> 
> So I don't really need this, I could roll my own in-place. I'm doing this because Francois could use it in some places where he uses the whole "write into a file then read it out". 

Is that code going to be in the Fuchsia tree or the Crashpad tree? That's the part I wasn't sure about.

Still it seems that anyone who needs to create a minidump with crashpad will need to do the whole read into the vmo.
> 
> Anyway, let me know if you don't want me to do this and I'll abandon the CL.

If it's going to be used in the Crashpad tree (in >= 1 places) then it's perfect to add it here.

Patch-set: 3