Update patch set 3
Patch Set 3:

To give a bit more context, the use case I have is for VMO to FileWriter for the attachments rather than the other way around. And looking at the latest patchset, the proposed functionality has been reduced to FileWriter to VMO, which I think only minidump producers need (and today that would only be the exception broker in the Fuchsia tree).

Patch-set: 3