Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Yes, I’m very very sorry about this. Over the past month, I’ve been absolutely slammed with mostly unrelated stuff, and just when I think I’m about to dig out, it starts up again. I will definitely come back to this soon.
> 
> In the meantime, if torne or tobiasjs have any comments (aside from remarks on my horrible tardiness), they wouldn’t be inappropriate.

I don't have any particularly specific comments about the code I think. I do still think that the encoding scheme that converts 9 bytes to 11 base94 symbols would be simpler/easier to understand (the one you linked to an example of when we were discussing it) than the 13/14 bits to 2 symbols scheme here, but as the discussion on Michael's design doc shows they perform pretty similarly and there isn't a strong overriding reason to pick one or the other.

Patch-set: 2