Update patch set 3
Patch Set 3:

(11 comments)

Ok, I'm happy that this is correct, although the fact that you can 'borrow' zero bits at the end of the stream is a little bit weird (but also quite clever).

Non-owner LGTM; it's your call how you deal with the comments below that came out of getting to the point where I understood the edge case correctness.

I think you should add

  {0xff, 0xff, 0xff, 0xff, 0x7f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}

as a test value, because that explicitly tests the case where there are 7 bits remaining after outputting an integral number of symbol pairs, but the final 7 bits can be encoded in half a symbol pair. It might be good to explain that that's what's happening in FinishDecoding.

Patch-set: 3