[Windows] Add checks for DLL loader lock.
This CL adds code to check if the current thread holds the DLL loader
lock. This code can be used to enforce the requirement that certain
pieces of crashpad, such as process creation and not done during calls
to DllMain which can lead to deadlocks and crashes. Only one check is
current enforced in client process creation, and only in debug builds.

Bug: crashpad: 316
Change-Id: I5757a264bbf28ce2ab88a0cd7ac9481e46428c17