Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Could we codify this by adding the following .gitattributes file to this repository?
> 
>     * text eol=lf

Darn, sorry about that, it's been a long time since I've needed to double check this and clearly my editor was defaulting to CRLF and I hadn't change my config for that repo. Ensuring the requirement for LF only would be good in .gitattributes or at least a passing mention in the developing crashpad page so others don't make the same mistake.

Patch-set: 2