Update patch set 8
Patch Set 8:

> > > Right, that’s what I remembered. And your doc shows solution A, where the compressed minidump at 95%ile was 67kB, which is fairly close to microdump’s 51kB while being substantially more valuable.

I can't recall an instance in which being able to see other threads helped me debug a crash, tbh.. so while having more information if possible is certainly good, I suspect the costs of including it might be too high. The bigger the dump is the less likely we are to actually receive it intact.

> > > This new patch implements solution B, but where was the consensus that this was necessary? Even the doc says that we’ll start with solution A first. (I just rechecked, it still says that.)

I was expecting that we would start with logging the *smallest* amount of information possible to maximise the chance that we receive it, and then look at the stats on the real world sizes to see if we could potentially afford to make it bigger. But, yes, the doc says otherwise..

> > > Now that solution A has been implemented, what’s the experience been?

We haven't enabled/shipped it AFAIK so I don't think we have any experience :)

> > > Has something changed since we discussed it that prompts scaling it back? It seems that we should discuss those concerns. As it is now, the well-written and reviewed design doc doesn’t address what this patch is doing.
> > 
> > Richard had concern about the full size see the comment in the doc (https://docs.google.com/document/d/11LBZ5ii3B4jaC6R2bxxSTONRejkQO1qP_racgsOQ9W0/edit?disco=AAAADOa1_5k), based on that, I directly go to the solution B, will update the design doc.
> 
> But this approach wasn’t even discussed in that comment thread! It ended with Torne talking about the module list, but no decision there either.

I suggested we could strip back the module list because the analysis of the size I was looking at said that was the largest component, and I knew that we had only included a subset of the module info in microdumps, so it seemed like a good additional opportunity. I imagine I assumed that only sending the main thread info was already known/on the table and so didn't explicitly mention it as an additional change.

As I said I was assuming we would start with the *smallest* possible dumps. The Android log buffers are a shared resource, many of the clients are very badly behaved, and the log output may not actually be read until some time after the original crash occurs in many cases, so the bigger our dumps are the less likely they are to be received intact. Even our own implementation can fight with itself; with microdumps it was not uncommon for a bunch of different chromium processes to crash around the same time and all write their dumps to the log at once, ending up interleaved and often incomplete.

The dump in the log also often obscures information for developers; anything logged before the dump may be so far back in scrollback that nobody sees it (we saw this often with microdumps, where effectively nobody *ever* read the Java exception stack printed to the logs if there was a subsequent native crash). Actually improving this problem would likely require that we make the dumps *much* smaller though (smaller than microdumps), so that might not really be feasible.

So, yeah - aiming for the rough size of microdumps is a reasonable guideline because that was generally working for us before, but I suspect that we could benefit from managing to dump something that's smaller; even average-size microdumps were already "much too big" from some perspectives.

Also: the doc talks about capping the size but seems to imply that dumps exceeding the cap will just not work (corrupt/truncated?), which seems like the worst option all round - if we can't "cleanly" terminate the dump when it's getting too big and still be able to decode key information (like the crashing thread's stack) then I don't think having a cap at all is desirable; there was no cap on the size of microdumps.

If we want to go ahead with enabling the implementation without trying to reduce the dump size any more first, then that's fine, but I believe that our goal here should be "what's the smallest number of bytes we can write to the log that give us enough info to make sense of most crashes", not "what's the most useful info we can get into a dump that's no larger than X", regardless of the value of X - and it seems like we're not in agreement there.

Patch-set: 8
Reviewer: Gerrit User 1109505 <1109505@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>